arch: arm: aarch32: Rename `cortex_r` to `cortex_ar`
This commit renames the `cortex_r` directory under the AArch32 to
`cortex_ar`, in preparation for the AArch32 Cortex-A support.

The rationale for this renaming is that the Cortex-A and Cortex-R share
the same base design and the difference between them, other than the
MPU vs. MMU, is minimal.

Since most of the architecture port code and configurations will be
shared between the Cortex-A and Cortex-R architectures, it is
advantageous to have them together in the same directory.
