+++ arch/arm/core/aarch32/CMakeLists.txt
@@ -32,6 +32,6 @@ add_subdirectory_ifdef(CONFIG_CPU_CORTEX_M_HAS_CMSE cortex_m/cmse)
 add_subdirectory_ifdef(CONFIG_ARM_SECURE_FIRMWARE cortex_m/tz)
 add_subdirectory_ifdef(CONFIG_ARM_NONSECURE_FIRMWARE cortex_m/tz)
 
-add_subdirectory_ifdef(CONFIG_CPU_CORTEX_R cortex_r)
+add_subdirectory_ifdef(CONFIG_CPU_CORTEX_R cortex_ar)
 
 zephyr_linker_sources(ROM_START SORT_KEY 0x0vectors vector_table.ld)
