+++ arch/arm/core/aarch32/Kconfig
@@ -238,6 +238,6 @@ config FP_SOFTABI
 endchoice
 
 rsource "cortex_m/Kconfig"
-rsource "cortex_r/Kconfig"
+rsource "cortex_ar/Kconfig"
 
 endif # !ARM64
