+++ arch/arm/core/aarch32/cortex_ar/fault.c
