+++ arch/arm/core/aarch32/cortex_ar/vector_table.h
