+++ include/arch/arm/aarch32/cortex_ar/scripts/app_data_alignment.ld
