+++ soc/arm/xilinx_zynqmp/linker.ld
@@ -8,5 +8,5 @@
 #include <autoconf.h>
 
 #if defined(CONFIG_SOC_XILINX_ZYNQMP_RPU)
-#include <arch/arm/aarch32/cortex_r/scripts/linker.ld>
+#include <arch/arm/aarch32/cortex_ar/scripts/linker.ld>
 #endif
