+++ subsys/testsuite/include/test_asm_inline_gcc.h
@@ -31,7 +31,7 @@ static inline void timestamp_serialize(void)
 	__ISB();
 }
 #elif defined(CONFIG_CPU_CORTEX_R)
-#include <arch/arm/aarch32/cortex_r/cpu.h>
+#include <arch/arm/aarch32/cortex_ar/cpu.h>
 static inline void timestamp_serialize(void)
 {
 	__ISB();
